# openapi (4).yaml
openapi: 3.0.1
info:
  title: Book Handler
  description: 'This is a BookHandler server. BookHandler is a platform for sharing and reading books and graphical novels that can be described as online library. This application should be especially useful for book enthusiasts that are looking for rare prints, not available in internet bookstores and on general auction sites and people interested in expanding their personal collections.'
  contact:
    email: 212829@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
  /publication:
    put:
      tags:
      - publication
      summary: Update an existing publication
      operationId: updatePublication
      parameters:
      - name: user
        in: header
        description: User info to check modification rights
        required: true
        explode: true
        schema:
            $ref: '#/components/schemas/User'
      requestBody:
        description: Publication object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publication'
        required: true
      responses:
        204:
          description: Successfully updated publication object
        404:
          description: Publication not found
          content: {}
        403:
          description: User does not have the rights to alter this publication
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
      - publication
      summary: Add a new publication to the service
      operationId: addPublication
      requestBody:
        description: Publication object that needs to be added to the service
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publication'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
      x-codegen-request-body-name: body
  /publication/findByGenre:
    get:
      tags:
      - publication
      summary: Find publications by genre
      description: Multiple genre values can be provided with comma separated strings
      operationId: findPublicationsByGenre
      parameters:
      - name: genre
        in: query
        description: Genre values that need to be considered for filter
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Invalid genre value
          content: {}
  /publication/findByAuthor:
    get:
      tags:
      - publication
      summary: Find publications by author's name
      description: Muliple authors names can be provided with comma separated strings. 
      operationId: findPublicationsByAuthor
      parameters:
      - name: authors
        in: query
        description: Authors to filter by
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
  /publication/findByReleaseYear:
    get:
      tags:
      - publication
      summary: Find publications by release year
      description: Muliple years integers can be provided with comma separated integers. 
      operationId: findPublicationsByYear
      parameters:
      - name: years
        in: query
        description: Years to filter by
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: integer
            format: int32
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
  /publication/{publicationId}:
    get:
      tags:
      - publication
      summary: Find publication by ID
      description: Returns a single publication
      operationId: getPublicationById
      parameters:
      - name: publicationId
        in: path
        description: ID of publication to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Publication not found
          content: {}
    delete:
      tags:
      - publication
      summary: Delete a publication
      operationId: deletePublication
      parameters:
      - name: user
        in: cookie
        schema:
          $ref: '#/components/schemas/User'
      - name: publicationId
        in: path
        description: Publication id to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Publication not found
          content: {}
        403:
          description: User does not have the rights to delete this publication
          content: {}
    put:
      tags:
      - publication
      summary: Update rating of an existing publication
      operationId: updatePublicationRating
      parameters:
      - name: publicationId
        in: path
        description: ID of publication to rate
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Publication rating
        content:
          application/json:
            schema:
              type: integer
              format: int32
        required: true
      responses:
        204:
          description: Successfully updated publication object
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Publication not found
          content: {}
        403:
          description: User does not have the rights to alter this publication
          content: {}
      x-codegen-request-body-name: body
...
components:
  schemas:
    Publication:
      required:
      - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        releaseYear:
          type: integer
          format: int32
        genre:
          type: array
          items: 
            $ref: '#/components/schemas/Genre'
        dateAdded:
          type: string
          format: date-time
        rating:
          type: number
          format: float
        photoUrls:
          type: array
          xml:
            name: photoUrl
            wrapped: true
          items:
            type: string
        addedBy:
          $ref: '#/components/schemas/User'
    Genre:
      type: object
      properties:
        name:
          type: string
    User:
      required:
      - username
      - email
      - password
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        userRole:
          type: string
          description: User Status
          enum: 
            - admin
            - member
        dateAdded:
          type: string
          format: date-time
...